safari-600.7-branch/LayoutTests/inspector-obsolete/http-tests/injected-script-discard-expected.txt

CONSOLE MESSAGE: line 1: Opening front-end for the first time
CONSOLE MESSAGE: line 1: Resolved object successfully
CONSOLE MESSAGE: line 1: Opening front-end second time
CONSOLE MESSAGE: line 1: Error resolving object
Tests that injected script is discarded upon front-end close.
